Enum conjure_oxide::ast::Expression

#[non_exhaustive]
pub enum Expression {

    Nothing,
    ConstantInt(i32),
    ConstantBool(bool),
    Reference(Name),
    Sum(Vec<Expression>),
    Not(Box<Expression>),
    Or(Vec<Expression>),
    And(Vec<Expression>),
    Eq(Box<Expression>, Box<Expression>),
    Neq(Box<Expression>, Box<Expression>),
    Geq(Box<Expression>, Box<Expression>),
    Leq(Box<Expression>, Box<Expression>),
    Gt(Box<Expression>, Box<Expression>),
    Lt(Box<Expression>, Box<Expression>),
    SumEq(Vec<Expression>, Box<Expression>),
    SumGeq(Vec<Expression>, Box<Expression>),
    SumLeq(Vec<Expression>, Box<Expression>),
    Ineq(Box<Expression>, Box<Expression>, Box<Expression>),

}

Solver Support

Minion

ConstantInt

Ineq

Reference

SumGeq

SumLeq

SAT

And

ConstantInt

Not

Or

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Nothing

Represents an empty expression

NB: we only expect this at the top level of a model (if there is no constraints)

ConstantInt(i32)

Supported by: Minion, SAT.

ConstantBool(bool)

Reference(Name)

Supported by: Minion.

Sum(Vec<Expression>)

Not(Box<Expression>)

Supported by: SAT.

Or(Vec<Expression>)

Supported by: SAT.

And(Vec<Expression>)

Supported by: SAT.

Eq(Box<Expression>, Box<Expression>)

Neq(Box<Expression>, Box<Expression>)

Geq(Box<Expression>, Box<Expression>)

Leq(Box<Expression>, Box<Expression>)

Gt(Box<Expression>, Box<Expression>)

Lt(Box<Expression>, Box<Expression>)

SumEq(Vec<Expression>, Box<Expression>)

SumGeq(Vec<Expression>, Box<Expression>)

Supported by: Minion.

SumLeq(Vec<Expression>, Box<Expression>)

Supported by: Minion.

Ineq(Box<Expression>, Box<Expression>, Box<Expression>)

Supported by: Minion.

Implementations

impl Expression

pub fn sub_expressions(&self) -> Option<Vec<&Expression>>

Returns a vector of references to the sub-expressions of the expression. If the expression is a primitive (variable, constant, etc.), returns None.

Note: If the expression is NOT MEANT TO have sub-expressions, this function will return None. Otherwise, it will return Some(Vec), where the Vec can be empty.

pub fn with_sub_expressions(&self, sub: Vec<&Expression>) -> Expression

Returns a clone of the same expression type with the given sub-expressions.