conjure_oxide
Function get_example_model_by_path

pub fn get_example_model_by_path(filepath: &str) -> Result<Model, Error>

Searches for an .essence file at the given filepath, then uses conjure to process it into astjson, and returns the parsed model.

Arguments

filepath - A string slice that holds the full file path

Returns

Function returns a Result<Value, anyhow::Error>, where Value is the parsed model