conjure_oxide::solver
Struct Solver

pub struct Solver<A, State = Init>
where
    A: SolverAdaptor,
    State: SolverState,
{ /* private fields */ }

An abstract representation of a constraints solver.

Solver provides a common interface for interacting with a constraint solver. It also abstracts over solver-specific datatypes, handling the translation to/from conjure_core::ast types for a model and its solutions.

Details of how a model is solved is specified by the SolverAdaptor. This includes: the underlying solver used, the translation of the model to a solver compatible form, how solutions are translated back to conjure_core::ast types, and how incremental solving is implemented. As such, there may be multiple SolverAdaptor implementations for a single underlying solver: e.g. one adaptor may give solutions in a representation close to the solvers, while another may attempt to rewrite it back into Essence.

Implementations

impl<Adaptor> Solver<Adaptor>
where Adaptor: SolverAdaptor,

pub fn new(solver_adaptor: Adaptor) -> Solver<Adaptor>

pub fn get_family(&self) -> SolverFamily

impl<A> Solver<A>
where A: SolverAdaptor,

pub fn load_model( self, model: Model, ) -> Result<Solver<A, ModelLoaded>, SolverError>

impl<A> Solver<A, ModelLoaded>
where A: SolverAdaptor,

pub fn solve( self, callback: Box<dyn Fn(HashMap<Name, Literal>) -> bool + Send>, ) -> Result<Solver<A, ExecutionSuccess>, SolverError>

pub fn solve_mut( self, callback: Box<dyn Fn(HashMap<Name, Literal>, Box<dyn ModelModifier>) -> bool + Send>, ) -> Result<Solver<A, ExecutionSuccess>, SolverError>

impl<A> Solver<A, ExecutionSuccess>
where A: SolverAdaptor,

pub fn stats(&self) -> SolverStats

pub fn save_stats_to_context(&self)

pub fn wall_time_s(&self) -> f64