Trait conjure_oxide::unstable::solver_interface::SolverAdaptor

pub trait SolverAdaptor: Sealed {
    type Model: Clone;
    type Solution: Clone;
    type Modifier: ModelModifier;

    // Required methods
    fn solve(
        &mut self,
        model: Self::Model,
        callback: fn(_: HashMap<String, String>) -> bool,
        _: Internal
    ) -> Result<ExecutionSuccess, ExecutionFailure>;
    fn solve_mut(
        &mut self,
        model: Self::Model,
        callback: fn(_: HashMap<String, String>, _: Self::Modifier) -> bool,
        _: Internal
    ) -> Result<ExecutionSuccess, ExecutionFailure>;
    fn load_model(
        &mut self,
        model: Model,
        _: Internal
    ) -> Result<Self::Model, Error>;

    // Provided method
    fn init_solver(&mut self, _: Internal) { ... }
}

A SolverAdaptor provide an interface to an underlying solver. Used by Solver.

Required Associated Types

type Model: Clone

The native model type of the underlying solver.

type Solution: Clone

The native solution type of the underlying solver.

type Modifier: ModelModifier

The ModelModifier used during incremental search.

If incremental solving is not supported, this SHOULD be set to NotModifiable .

Required Methods

fn solve( &mut self, model: Self::Model, callback: fn(_: HashMap<String, String>) -> bool, _: Internal ) -> Result<ExecutionSuccess, ExecutionFailure>

Run the solver on the given model.

Implementations of this function MUST call the user provided callback whenever a solution is found. If the user callback returns true, search should continue, if the user callback returns false, search should terminate.

fn solve_mut( &mut self, model: Self::Model, callback: fn(_: HashMap<String, String>, _: Self::Modifier) -> bool, _: Internal ) -> Result<ExecutionSuccess, ExecutionFailure>

Run the solver on the given model, allowing modification of the model through a ModelModifier.

Implementations of this function MUST return OpNotSupported if modifying the model mid-search is not supported. These implementations may also find the NotModifiable modifier useful.

As with solve, this function MUST call the user provided callback function whenever a solution is found.

fn load_model( &mut self, model: Model, _: Internal ) -> Result<Self::Model, Error>

Provided Methods

fn init_solver(&mut self, _: Internal)