conjure_oxide::ast
Enum Expression

pub enum Expression {

    Nothing,
    Bubble(Metadata, Box<Expression>, Box<Expression>),
    Constant(Metadata, Constant),
    Reference(Metadata, Name),
    Sum(Metadata, Vec<Expression>),
    Min(Metadata, Vec<Expression>),
    Max(Metadata, Vec<Expression>),
    Not(Metadata, Box<Expression>),
    Or(Metadata, Vec<Expression>),
    And(Metadata, Vec<Expression>),
    Eq(Metadata, Box<Expression>, Box<Expression>),
    Neq(Metadata, Box<Expression>, Box<Expression>),
    Geq(Metadata, Box<Expression>, Box<Expression>),
    Leq(Metadata, Box<Expression>, Box<Expression>),
    Gt(Metadata, Box<Expression>, Box<Expression>),
    Lt(Metadata, Box<Expression>, Box<Expression>),
    SafeDiv(Metadata, Box<Expression>, Box<Expression>),
    UnsafeDiv(Metadata, Box<Expression>, Box<Expression>),
    SumEq(Metadata, Vec<Expression>, Box<Expression>),
    SumGeq(Metadata, Vec<Expression>, Box<Expression>),
    SumLeq(Metadata, Vec<Expression>, Box<Expression>),
    DivEq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>),
    Ineq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>),
    AllDiff(Metadata, Vec<Expression>),
    WatchedLiteral(Metadata, Name, Constant),
    Reify(Metadata, Box<Expression>, Box<Expression>),

}

Represents different types of expressions used to define rules and constraints in the model.

The Expression enum includes operations, constants, and variable references used to build rules and conditions for the model.

Compatability

JsonInput

And

Constant

Eq

Geq

Gt

Leq

Lt

Max

Min

Neq

Not

Or

Reference

Sum

UnsafeDiv

SAT

And

Not

Or

Reference

Minion

AllDiff

Constant

DivEq

Ineq

Reference

Reify

Sum

SumGeq

SumLeq

WatchedLiteral

Variants

Nothing

Represents an empty expression

NB: we only expect this at the top level of a model (if there is no constraints)

Bubble(Metadata, Box<Expression>, Box<Expression>)

An expression representing “A is valid as long as B is true” Turns into a conjunction when it reaches a boolean context

Constant(Metadata, Constant)

Supported by: Minion, JsonInput.

Reference(Metadata, Name)

Supported by: Minion, JsonInput, SAT.

Sum(Metadata, Vec<Expression>)

Supported by: Minion, JsonInput.

Min(Metadata, Vec<Expression>)

Supported by: JsonInput.

Max(Metadata, Vec<Expression>)

Supported by: JsonInput.

Not(Metadata, Box<Expression>)

Supported by: JsonInput, SAT.

Or(Metadata, Vec<Expression>)

Supported by: JsonInput, SAT.

And(Metadata, Vec<Expression>)

Supported by: JsonInput, SAT.

Eq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Neq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Geq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Leq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Gt(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Lt(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

SafeDiv(Metadata, Box<Expression>, Box<Expression>)

Division after preventing division by zero, usually with a bubble

UnsafeDiv(Metadata, Box<Expression>, Box<Expression>)

Division with a possibly undefined value (division by 0) Supported by: JsonInput.

SumEq(Metadata, Vec<Expression>, Box<Expression>)

SumGeq(Metadata, Vec<Expression>, Box<Expression>)

Supported by: Minion.

SumLeq(Metadata, Vec<Expression>, Box<Expression>)

Supported by: Minion.

DivEq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>)

Supported by: Minion.

Ineq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>)

Supported by: Minion.

AllDiff(Metadata, Vec<Expression>)

Supported by: Minion.

WatchedLiteral(Metadata, Name, Constant)

w-literal(x,k) is SAT iff x == k, where x is a variable and k a constant.

This is a low-level Minion constraint and you should (probably) use Eq instead. The main use of w-literal is to convert boolean variables to constraints so that they can be used inside watched-and and watched-or.

See rules::minion::boolean_literal_to_wliteral.

Supported by: Minion.

Reify(Metadata, Box<Expression>, Box<Expression>)

Supported by: Minion.

Implementations

impl Expression

pub fn domain_of( &self, vars: &HashMap<Name, DecisionVariable>, ) -> Option<Domain>

Returns the possible values of the expression, recursing to leaf expressions

pub fn can_be_undefined(&self) -> bool

pub fn return_type(&self) -> Option<ReturnType>

pub fn is_clean(&self) -> bool

pub fn set_clean(&mut self, bool_value: bool)