Enum conjure_oxide::ast::Expression

#[non_exhaustive]
pub enum Expression {

    Nothing,
    Bubble(Metadata, Box<Expression>, Box<Expression>),
    Constant(Metadata, Constant),
    Reference(Metadata, Name),
    Sum(Metadata, Vec<Expression>),
    Min(Metadata, Vec<Expression>),
    Not(Metadata, Box<Expression>),
    Or(Metadata, Vec<Expression>),
    And(Metadata, Vec<Expression>),
    Eq(Metadata, Box<Expression>, Box<Expression>),
    Neq(Metadata, Box<Expression>, Box<Expression>),
    Geq(Metadata, Box<Expression>, Box<Expression>),
    Leq(Metadata, Box<Expression>, Box<Expression>),
    Gt(Metadata, Box<Expression>, Box<Expression>),
    Lt(Metadata, Box<Expression>, Box<Expression>),
    SafeDiv(Metadata, Box<Expression>, Box<Expression>),
    UnsafeDiv(Metadata, Box<Expression>, Box<Expression>),
    SumEq(Metadata, Vec<Expression>, Box<Expression>),
    SumGeq(Metadata, Vec<Expression>, Box<Expression>),
    SumLeq(Metadata, Vec<Expression>, Box<Expression>),
    DivEq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>),
    Ineq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>),
    AllDiff(Metadata, Vec<Expression>),

}

Compatability

JsonInput

And

Constant

Eq

Geq

Gt

Leq

Lt

Min

Neq

Not

Or

Reference

Sum

UnsafeDiv

SAT

And

Not

Or

Reference

Minion

AllDiff

Constant

DivEq

Ineq

Reference

Sum

SumGeq

SumLeq

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Nothing

Represents an empty expression

NB: we only expect this at the top level of a model (if there is no constraints)

Bubble(Metadata, Box<Expression>, Box<Expression>)

An expression representing “A is valid as long as B is true” Turns into a conjunction when it reaches a boolean context

Constant(Metadata, Constant)

Supported by: Minion, JsonInput.

Reference(Metadata, Name)

Supported by: Minion, JsonInput, SAT.

Sum(Metadata, Vec<Expression>)

Supported by: Minion, JsonInput.

Min(Metadata, Vec<Expression>)

Supported by: JsonInput.

Not(Metadata, Box<Expression>)

Supported by: JsonInput, SAT.

Or(Metadata, Vec<Expression>)

Supported by: JsonInput, SAT.

And(Metadata, Vec<Expression>)

Supported by: JsonInput, SAT.

Eq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Neq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Geq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Leq(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Gt(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

Lt(Metadata, Box<Expression>, Box<Expression>)

Supported by: JsonInput.

SafeDiv(Metadata, Box<Expression>, Box<Expression>)

Division after preventing division by zero, usually with a bubble

UnsafeDiv(Metadata, Box<Expression>, Box<Expression>)

Division with a possibly undefined value (division by 0) Supported by: JsonInput.

SumEq(Metadata, Vec<Expression>, Box<Expression>)

SumGeq(Metadata, Vec<Expression>, Box<Expression>)

Supported by: Minion.

SumLeq(Metadata, Vec<Expression>, Box<Expression>)

Supported by: Minion.

DivEq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>)

Supported by: Minion.

Ineq(Metadata, Box<Expression>, Box<Expression>, Box<Expression>)

Supported by: Minion.

AllDiff(Metadata, Vec<Expression>)

Supported by: Minion.

Implementations

impl Expression

enum Variant Predicates

pub fn is_nothing(&self) -> bool

Is this Expression a Nothing?

pub fn is_bubble(&self) -> bool

Is this Expression a Bubble?

pub fn is_constant(&self) -> bool

Is this Expression a Constant?

pub fn is_reference(&self) -> bool

Is this Expression a Reference?

pub fn is_sum(&self) -> bool

Is this Expression a Sum?

pub fn is_min(&self) -> bool

Is this Expression a Min?

pub fn is_not(&self) -> bool

Is this Expression a Not?

pub fn is_or(&self) -> bool

Is this Expression a Or?

pub fn is_and(&self) -> bool

Is this Expression a And?

pub fn is_eq(&self) -> bool

Is this Expression a Eq?

pub fn is_neq(&self) -> bool

Is this Expression a Neq?

pub fn is_geq(&self) -> bool

Is this Expression a Geq?

pub fn is_leq(&self) -> bool

Is this Expression a Leq?

pub fn is_gt(&self) -> bool

Is this Expression a Gt?

pub fn is_lt(&self) -> bool

Is this Expression a Lt?

pub fn is_safe_div(&self) -> bool

Is this Expression a SafeDiv?

pub fn is_unsafe_div(&self) -> bool

Is this Expression a UnsafeDiv?

pub fn is_sum_eq(&self) -> bool

Is this Expression a SumEq?

pub fn is_sum_geq(&self) -> bool

Is this Expression a SumGeq?

pub fn is_sum_leq(&self) -> bool

Is this Expression a SumLeq?

pub fn is_div_eq(&self) -> bool

Is this Expression a DivEq?

pub fn is_ineq(&self) -> bool

Is this Expression a Ineq?

pub fn is_all_diff(&self) -> bool

Is this Expression a AllDiff?

impl Expression

pub fn domain_of( &self, vars: &HashMap<Name, DecisionVariable> ) -> Option<Domain>

Returns the possible values of the expression, recursing to leaf expressions

pub fn can_be_undefined(&self) -> bool

pub fn return_type(&self) -> Option<ReturnType>

pub fn is_clean(&self) -> bool

pub fn set_clean(&mut self, bool_value: bool)